split message generation
