+++ ros/vilib_msgs/.clang-format
@@ -1,0 +1,2 @@
+DisableFormat: true
+SortIncludes: false
