+++ ros/vilib_msgs/package.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0"?>
 <package format="2">
-  <name>vilib</name>
+  <name>vilib_msgs</name>
   <version>1.0.0</version>
-  <description>Fast CUDA Feature Tracker</description>
+  <description>vilib message definitions</description>
 
   <maintainer email="foehn@ifi.uzh.ch">foehnph</maintainer>
 
@@ -42,9 +42,5 @@
   <buildtool_depend>catkin</buildtool_depend>
   <buildtool_depend>catkin_simple</buildtool_depend>
   <depend>message_generation</depend>
-  <depend>cv_bridge</depend>
-  <depend>image_transport</depend>
-  <depend>sensor_msgs</depend>
-  <depend>geometry_msgs</depend>
-  <depend>roscpp</depend>
+  <depend>std_msgs</depend>
 </package>
