+++ ros/vilib_tracker/include/vilib_tracker/logger.hpp
